Refactor: 파일 위치 변경, axios 함수 반환 제네릭 타입 설정
