+++ src/api/league-game/index.ts
@@ -1,11 +1,11 @@
-import useGetLeagueGameAll from './use-get-league-game-all';
-import useGetLeagueGameByDate from './use-get-league-game-by-date';
-import useGetLeagueGameByLeagueGameId from './use-get-league-game-by-league-game-id';
-import useGetLeagueGameByYearMonth from './use-get-league-game-by-year-month';
+import useGetAllGameList from './use-get-all-game-list';
+import useGetGameListByDate from './use-get-game-list-by-date';
+import useGetGameByLeagueGameId from './use-get-game-by-league-game-id';
+import useGetGameListByLeagueGameIdYearMonth from './use-get-game-list-by-league-game-id-year-month';
 
 export {
-  useGetLeagueGameAll,
-  useGetLeagueGameByDate,
-  useGetLeagueGameByYearMonth,
-  useGetLeagueGameByLeagueGameId,
+  useGetAllGameList,
+  useGetGameListByDate,
+  useGetGameListByLeagueGameIdYearMonth,
+  useGetGameByLeagueGameId,
 };
