+++ src/api/league-game/league-games.ts
@@ -1,35 +1,45 @@
-import { instance } from '@/utils/intance';
+import { instance } from '@/api/intance';
+import {
+  TAllGameList,
+  TGameByLeagueGameId,
+  TGameListByDate,
+  TGameListByLeagueIdYearMonth,
+} from '@/types/league-games';
 
 const ENDPOINT = '/leagueGame';
 
-export async function getLeagueGameByLeagueGameId(leagueGameId: number) {
-  const { data } = await instance.get(`${ENDPOINT}/${leagueGameId}`);
+export async function getGameByLeagueGameId(leagueGameId: number) {
+  const { data } = await instance.get<TGameByLeagueGameId>(`${ENDPOINT}/${leagueGameId}`);
 
-  return data;
+  return data.response;
 }
 
-export async function getLeagueGameByYearMonth(leagueId: number, yearMonth: string) {
-  const { data } = await instance.get(`${ENDPOINT}/${leagueId}/${yearMonth}`);
+export async function getGameListByLeagueIdYearMonth(leagueId: number, yearMonth: string) {
+  const { data } = await instance.get<TGameListByLeagueIdYearMonth>(
+    `${ENDPOINT}/${leagueId}/${yearMonth}`,
+  );
 
-  return data;
+  return data.responses;
 }
 
-export async function getLeagueGameByDate(targetDate: string) {
-  const { data } = await instance.get(`${ENDPOINT}/date/${targetDate}`);
+export async function getGameListByDate(targetDate: string) {
+  const { data } = await instance.get<TGameListByDate>(`${ENDPOINT}/date/${targetDate}`);
 
-  return data;
+  return data.responses;
 }
 
-export async function getLeagueGameAll(
+export async function getAllGameList(
   startDate: string,
   endDate: string,
   leagueId: number,
   page?: number,
   size?: number,
 ) {
-  const { data } = await instance.get(`${ENDPOINT}/all`, {
+  const { data } = await instance.get<TAllGameList>(`${ENDPOINT}/all`, {
     params: { startDate, endDate, leagueId, page, size },
   });
 
-  return data;
+  const { findLeagueGames, totalElements, totalPages } = data;
+
+  return { findLeagueGames, totalElements, totalPages };
 }
